Change how the file is ingested when ingesting a bag. This seems to correct the problems I was seeing before.
At this point a collection object can be created from a bag. This
is probably not the final form that we want, but it is a step.

The next step may be to retrench a bit. I'd like to add some tests
and to add proper namespacing. Better to do that now than wait too much
longer.